# client/openapi.json
{"openapi": "3.1.0", "info": {"title": "NinjaExtraAPI", "version": "1.0.0", "description": ""}, "paths": {"/api/token/pair": {"post": {"operationId": "token_obtain_pair", "summary": "Obtain Token", "parameters": [], "responses": {"200": {"description": "OK", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/TokenObtainPairOutputSchema"}}}}, "422": {"description": "Unprocessable Entity", "content": {"application/json": {"schema": {"type": "object", "properties": {"tag": {"type": "string", "const": "ValidationError"}, "detail": {"type": "string"}, "errors": {"type": "array", "items": {"type": "object", "properties": {"type": {"type": "string"}, "loc": {"type": "array", "items": {"type": "string"}}, "msg": {"type": "string"}, "ctx": {"type": "object"}}, "required": ["type", "loc", "msg"]}}}, "required": ["tag", "detail", "errors"], "title": "ValidationError"}}}}, "500": {"description": "Internal Server Error", "content": {"application/json": {"schema": {"type": "object", "properties": {"tag": {"type": "string", "const": "InternalError"}, "detail": {"type": "string"}}, "required": ["tag", "detail"], "title": "InternalError"}}}}}, "tags": ["token"], "requestBody": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/TokenObtainPairInputSchema"}}}, "required": true}}}, "/api/token/refresh": {"post": {"operationId": "token_refresh", "summary": "Refresh Token", "parameters": [], "responses": {"200": {"description": "OK", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/TokenRefreshOutputSchema"}}}}, "422": {"description": "Unprocessable Entity", "content": {"application/json": {"schema": {"type": "object", "properties": {"tag": {"type": "string", "const": "ValidationError"}, "detail": {"type": "string"}, "errors": {"type": "array", "items": {"type": "object", "properties": {"type": {"type": "string"}, "loc": {"type": "array", "items": {"type": "string"}}, "msg": {"type": "string"}, "ctx": {"type": "object"}}, "required": ["type", "loc", "msg"]}}}, "required": ["tag", "detail", "errors"], "title": "ValidationError"}}}}, "500": {"description": "Internal Server Error", "content": {"application/json": {"schema": {"type": "object", "properties": {"tag": {"type": "string", "const": "InternalError"}, "detail": {"type": "string"}}, "required": ["tag", "detail"], "title": "InternalError"}}}}}, "tags": ["token"], "requestBody": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/TokenRefreshInputSchema"}}}, "required": true}}}, "/api/token/verify": {"post": {"operationId": "token_verify", "summary": "Verify Token", "parameters": [], "responses": {"200": {"description": "OK", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Schema"}}}}, "422": {"description": "Unprocessable Entity", "content": {"application/json": {"schema": {"type": "object", "properties": {"tag": {"type": "string", "const": "ValidationError"}, "detail": {"type": "string"}, "errors": {"type": "array", "items": {"type": "object", "properties": {"type": {"type": "string"}, "loc": {"type": "array", "items": {"type": "string"}}, "msg": {"type": "string"}, "ctx": {"type": "object"}}, "required": ["type", "loc", "msg"]}}}, "required": ["tag", "detail", "errors"], "title": "ValidationError"}}}}, "500": {"description": "Internal Server Error", "content": {"application/json": {"schema": {"type": "object", "properties": {"tag": {"type": "string", "const": "InternalError"}, "detail": {"type": "string"}}, "required": ["tag", "detail"], "title": "InternalError"}}}}}, "tags": ["token"], "requestBody": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/TokenVerifyInputSchema"}}}, "required": true}}}, "/api/blog/posts/": {"get": {"operationId": "blog_api_get_posts", "summary": "Get Posts", "parameters": [], "responses": {"200": {"description": "OK", "content": {"application/json": {"schema": {"items": {"$ref": "#/components/schemas/PostResponse"}, "title": "Response", "type": "array"}}}}, "422": {"description": "Unprocessable Entity", "content": {"application/json": {"schema": {"type": "object", "properties": {"tag": {"type": "string", "const": "ValidationError"}, "detail": {"type": "string"}, "errors": {"type": "array", "items": {"type": "object", "properties": {"type": {"type": "string"}, "loc": {"type": "array", "items": {"type": "string"}}, "msg": {"type": "string"}, "ctx": {"type": "object"}}, "required": ["type", "loc", "msg"]}}}, "required": ["tag", "detail", "errors"], "title": "ValidationError"}}}}, "500": {"description": "Internal Server Error", "content": {"application/json": {"schema": {"type": "object", "properties": {"tag": {"type": "string", "const": "InternalError"}, "detail": {"type": "string"}}, "required": ["tag", "detail"], "title": "InternalError"}}}}}}}, "/api/blog/post/{post_id}": {"get": {"operationId": "blog_api_get_post", "summary": "Get Post", "parameters": [{"in": "path", "name": "post_id", "schema": {"title": "Post Id", "type": "integer"}, "required": true}], "responses": {"200": {"description": "OK", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/PostResponse"}}}}, "400": {"description": "Bad Request", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/PostErrors"}}}}, "422": {"description": "Unprocessable Entity", "content": {"application/json": {"schema": {"type": "object", "properties": {"tag": {"type": "string", "const": "ValidationError"}, "detail": {"type": "string"}, "errors": {"type": "array", "items": {"type": "object", "properties": {"type": {"type": "string"}, "loc": {"type": "array", "items": {"type": "string"}}, "msg": {"type": "string"}, "ctx": {"type": "object"}}, "required": ["type", "loc", "msg"]}}}, "required": ["tag", "detail", "errors"], "title": "ValidationError"}}}}, "500": {"description": "Internal Server Error", "content": {"application/json": {"schema": {"type": "object", "properties": {"tag": {"type": "string", "const": "InternalError"}, "detail": {"type": "string"}}, "required": ["tag", "detail"], "title": "InternalError"}}}}}}}, "/api/orders/": {"get": {"operationId": "orders_api_list_orders", "summary": "List Orders", "parameters": [{"in": "query", "name": "status", "schema": {"anyOf": [{"$ref": "#/components/schemas/Status"}, {"type": "null"}]}, "required": false}, {"in": "query", "name": "q", "schema": {"anyOf": [{"type": "string"}, {"type": "null"}], "q": ["customer_name__icontains"], "title": "Q"}, "required": false}, {"in": "query", "name": "min_total", "schema": {"anyOf": [{"type": "number"}, {"pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$", "type": "string"}, {"type": "null"}], "q": "total_price__gte", "title": "Min Total"}, "required": false}, {"in": "query", "name": "max_total", "schema": {"anyOf": [{"type": "number"}, {"pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$", "type": "string"}, {"type": "null"}], "q": "total_price__lte", "title": "Max Total"}, "required": false}], "responses": {"200": {"description": "OK", "content": {"application/json": {"schema": {"items": {"$ref": "#/components/schemas/OrderResponse"}, "title": "Response", "type": "array"}}}}, "422": {"description": "Unprocessable Entity", "content": {"application/json": {"schema": {"type": "object", "properties": {"tag": {"type": "string", "const": "ValidationError"}, "detail": {"type": "string"}, "errors": {"type": "array", "items": {"type": "object", "properties": {"type": {"type": "string"}, "loc": {"type": "array", "items": {"type": "string"}}, "msg": {"type": "string"}, "ctx": {"type": "object"}}, "required": ["type", "loc", "msg"]}}}, "required": ["tag", "detail", "errors"], "title": "ValidationError"}}}}, "500": {"description": "Internal Server Error", "content": {"application/json": {"schema": {"type": "object", "properties": {"tag": {"type": "string", "const": "InternalError"}, "detail": {"type": "string"}}, "required": ["tag", "detail"], "title": "InternalError"}}}}, "401": {"description": "Unauthorized", "content": {"application/json": {"schema": {"type": "object", "properties": {"tag": {"type": "string", "const": "AuthenticationError"}, "detail": {"type": "string"}}, "required": ["tag", "detail"], "title": "AuthenticationError"}}}}, "403": {"description": "Forbidden", "content": {"application/json": {"schema": {"type": "object", "properties": {"tag": {"type": "string", "const": "AuthorizationError"}, "detail": {"type": "string"}}, "required": ["tag", "detail"], "title": "AuthorizationError"}}}}}, "security": [{"JWTAuth": []}]}}, "/api/orders/{order_id}/": {"get": {"operationId": "orders_api_get_order", "summary": "Get Order", "parameters": [{"in": "path", "name": "order_id", "schema": {"title": "Order Id", "type": "integer"}, "required": true}], "responses": {"200": {"description": "OK", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/OrderResponse"}}}}, "400": {"description": "Bad Request", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/OrderErrors"}}}}, "422": {"description": "Unprocessable Entity", "content": {"application/json": {"schema": {"type": "object", "properties": {"tag": {"type": "string", "const": "ValidationError"}, "detail": {"type": "string"}, "errors": {"type": "array", "items": {"type": "object", "properties": {"type": {"type": "string"}, "loc": {"type": "array", "items": {"type": "string"}}, "msg": {"type": "string"}, "ctx": {"type": "object"}}, "required": ["type", "loc", "msg"]}}}, "required": ["tag", "detail", "errors"], "title": "ValidationError"}}}}, "500": {"description": "Internal Server Error", "content": {"application/json": {"schema": {"type": "object", "properties": {"tag": {"type": "string", "const": "InternalError"}, "detail": {"type": "string"}}, "required": ["tag", "detail"], "title": "InternalError"}}}}, "401": {"description": "Unauthorized", "content": {"application/json": {"schema": {"type": "object", "properties": {"tag": {"type": "string", "const": "AuthenticationError"}, "detail": {"type": "string"}}, "required": ["tag", "detail"], "title": "AuthenticationError"}}}}, "403": {"description": "Forbidden", "content": {"application/json": {"schema": {"type": "object", "properties": {"tag": {"type": "string", "const": "AuthorizationError"}, "detail": {"type": "string"}}, "required": ["tag", "detail"], "title": "AuthorizationError"}}}}}, "security": [{"JWTAuth": []}]}}, "/api/products/": {"get": {"operationId": "products_api_list_products", "summary": "List Products", "parameters": [], "responses": {"200": {"description": "OK", "content": {"application/json": {"schema": {"items": {"$ref": "#/components/schemas/ProductResponse"}, "title": "Response", "type": "array"}}}}, "422": {"description": "Unprocessable Entity", "content": {"application/json": {"schema": {"type": "object", "properties": {"tag": {"type": "string", "const": "ValidationError"}, "detail": {"type": "string"}, "errors": {"type": "array", "items": {"type": "object", "properties": {"type": {"type": "string"}, "loc": {"type": "array", "items": {"type": "string"}}, "msg": {"type": "string"}, "ctx": {"type": "object"}}, "required": ["type", "loc", "msg"]}}}, "required": ["tag", "detail", "errors"], "title": "ValidationError"}}}}, "500": {"description": "Internal Server Error", "content": {"application/json": {"schema": {"type": "object", "properties": {"tag": {"type": "string", "const": "InternalError"}, "detail": {"type": "string"}}, "required": ["tag", "detail"], "title": "InternalError"}}}}, "401": {"description": "Unauthorized", "content": {"application/json": {"schema": {"type": "object", "properties": {"tag": {"type": "string", "const": "AuthenticationError"}, "detail": {"type": "string"}}, "required": ["tag", "detail"], "title": "AuthenticationError"}}}}, "403": {"description": "Forbidden", "content": {"application/json": {"schema": {"type": "object", "properties": {"tag": {"type": "string", "const": "AuthorizationError"}, "detail": {"type": "string"}}, "required": ["tag", "detail"], "title": "AuthorizationError"}}}}}, "security": [{"JWTAuth": []}]}}, "/api/products/{product_id}/": {"get": {"operationId": "products_api_get_product", "summary": "Get Product", "parameters": [{"in": "path", "name": "product_id", "schema": {"title": "Product Id", "type": "integer"}, "required": true}], "responses": {"200": {"description": "OK", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ProductResponse"}}}}, "400": {"description": "Bad Request", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ProductErrors"}}}}, "422": {"description": "Unprocessable Entity", "content": {"application/json": {"schema": {"type": "object", "properties": {"tag": {"type": "string", "const": "ValidationError"}, "detail": {"type": "string"}, "errors": {"type": "array", "items": {"type": "object", "properties": {"type": {"type": "string"}, "loc": {"type": "array", "items": {"type": "string"}}, "msg": {"type": "string"}, "ctx": {"type": "object"}}, "required": ["type", "loc", "msg"]}}}, "required": ["tag", "detail", "errors"], "title": "ValidationError"}}}}, "500": {"description": "Internal Server Error", "content": {"application/json": {"schema": {"type": "object", "properties": {"tag": {"type": "string", "const": "InternalError"}, "detail": {"type": "string"}}, "required": ["tag", "detail"], "title": "InternalError"}}}}, "401": {"description": "Unauthorized", "content": {"application/json": {"schema": {"type": "object", "properties": {"tag": {"type": "string", "const": "AuthenticationError"}, "detail": {"type": "string"}}, "required": ["tag", "detail"], "title": "AuthenticationError"}}}}, "403": {"description": "Forbidden", "content": {"application/json": {"schema": {"type": "object", "properties": {"tag": {"type": "string", "const": "AuthorizationError"}, "detail": {"type": "string"}}, "required": ["tag", "detail"], "title": "AuthorizationError"}}}}}, "security": [{"JWTAuth": []}]}}}, "components": {"schemas": {"TokenObtainPairOutputSchema": {"properties": {"username": {"description": "Required. 150 characters or fewer. Letters, digits and @/./+/-/_ only.", "maxLength": 150, "title": "Username", "type": "string"}, "refresh": {"title": "Refresh", "type": "string"}, "access": {"title": "Access", "type": "string"}}, "required": ["username", "refresh", "access"], "title": "TokenObtainPairOutputSchema", "type": "object"}, "TokenObtainPairInputSchema": {"properties": {"password": {"format": "password", "title": "Password", "type": "string", "writeOnly": true}, "username": {"description": "Required. 150 characters or fewer. Letters, digits and @/./+/-/_ only.", "maxLength": 150, "title": "Username", "type": "string"}}, "required": ["password", "username"], "title": "TokenObtainPairInputSchema", "type": "object"}, "TokenRefreshOutputSchema": {"properties": {"refresh": {"title": "Refresh", "type": "string"}, "access": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Access"}}, "required": ["refresh", "access"], "title": "TokenRefreshOutputSchema", "type": "object"}, "TokenRefreshInputSchema": {"properties": {"refresh": {"title": "Refresh", "type": "string"}}, "required": ["refresh"], "title": "TokenRefreshInputSchema", "type": "object"}, "Schema": {"properties": {}, "title": "Schema", "type": "object"}, "TokenVerifyInputSchema": {"properties": {"token": {"title": "Token", "type": "string"}}, "required": ["token"], "title": "TokenVerifyInputSchema", "type": "object"}, "DraftPost": {"properties": {"tag": {"const": "draft", "default": "draft", "title": "Tag", "type": "string"}, "id": {"anyOf": [{"type": "integer"}, {"type": "null"}], "title": "ID"}, "title": {"maxLength": 250, "title": "Title", "type": "string"}, "slug": {"maxLength": 250, "title": "Slug", "type": "string"}, "updated": {"format": "date-time", "title": "Updated", "type": "string"}}, "required": ["title", "slug", "updated", "tag"], "title": "DraftPost", "type": "object"}, "PostResponse": {"anyOf": [{"$ref": "#/components/schemas/DraftPost"}, {"$ref": "#/components/schemas/PublishedPost"}], "title": "PostResponse"}, "PublishedPost": {"properties": {"tag": {"const": "published", "default": "published", "title": "Tag", "type": "string"}, "id": {"anyOf": [{"type": "integer"}, {"type": "null"}], "title": "ID"}, "title": {"maxLength": 250, "title": "Title", "type": "string"}, "slug": {"maxLength": 250, "title": "Slug", "type": "string"}, "body": {"title": "Body", "type": "string"}, "publish": {"format": "date-time", "title": "Publish", "type": "string"}}, "required": ["title", "slug", "body", "tag"], "title": "PublishedPost", "type": "object"}, "PostErrors": {"$ref": "#/components/schemas/PostNotFoundError", "title": "PostErrors"}, "PostNotFoundError": {"properties": {"tag": {"const": "PostNotFoundError", "default": "PostNotFoundError", "title": "Tag", "type": "string"}, "detail": {"title": "Detail", "type": "string"}, "id": {"title": "Id", "type": "integer"}}, "required": ["detail", "id", "tag"], "title": "PostNotFoundError", "type": "object"}, "OrderFilters": {"properties": {"status": {"anyOf": [{"$ref": "#/components/schemas/Status"}, {"type": "null"}]}, "q": {"anyOf": [{"type": "string"}, {"type": "null"}], "q": ["customer_name__icontains"], "title": "Q"}, "min_total": {"anyOf": [{"type": "number"}, {"pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$", "type": "string"}, {"type": "null"}], "q": "total_price__gte", "title": "Min Total"}, "max_total": {"anyOf": [{"type": "number"}, {"pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$", "type": "string"}, {"type": "null"}], "q": "total_price__lte", "title": "Max Total"}}, "title": "OrderFilters", "type": "object"}, "Status": {"enum": ["draft", "pending", "shipped", "cancelled"], "title": "Status", "type": "string"}, "CancelledOrder": {"properties": {"tag": {"const": "CancelledOrder", "default": "CancelledOrder", "title": "Tag", "type": "string"}, "id": {"title": "Id", "type": "integer"}, "customer_name": {"title": "Customer Name", "type": "string"}, "items_count": {"title": "Items Count", "type": "integer"}, "total_price": {"pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$", "title": "Total Price", "type": "string"}, "cancellation_reason": {"title": "Cancellation Reason", "type": "string"}, "cancelled_at": {"format": "date-time", "title": "Cancelled At", "type": "string"}, "created_at": {"format": "date-time", "title": "Created At", "type": "string"}}, "required": ["id", "customer_name", "items_count", "total_price", "cancellation_reason", "cancelled_at", "created_at", "tag"], "title": "CancelledOrder", "type": "object"}, "OrderResponse": {"anyOf": [{"$ref": "#/components/schemas/PendingOrder"}, {"$ref": "#/components/schemas/ShippedOrder"}, {"$ref": "#/components/schemas/CancelledOrder"}], "title": "OrderResponse"}, "PendingOrder": {"properties": {"tag": {"const": "PendingOrder", "default": "PendingOrder", "title": "Tag", "type": "string"}, "id": {"title": "Id", "type": "integer"}, "customer_name": {"title": "Customer Name", "type": "string"}, "items_count": {"title": "Items Count", "type": "integer"}, "total_price": {"pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$", "title": "Total Price", "type": "string"}, "created_at": {"format": "date-time", "title": "Created At", "type": "string"}}, "required": ["id", "customer_name", "items_count", "total_price", "created_at", "tag"], "title": "PendingOrder", "type": "object"}, "ShippedOrder": {"properties": {"tag": {"const": "ShippedOrder", "default": "ShippedOrder", "title": "Tag", "type": "string"}, "id": {"title": "Id", "type": "integer"}, "customer_name": {"title": "Customer Name", "type": "string"}, "items_count": {"title": "Items Count", "type": "integer"}, "total_price": {"pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$", "title": "Total Price", "type": "string"}, "tracking_number": {"title": "Tracking Number", "type": "string"}, "shipped_at": {"format": "date-time", "title": "Shipped At", "type": "string"}, "created_at": {"format": "date-time", "title": "Created At", "type": "string"}}, "required": ["id", "customer_name", "items_count", "total_price", "tracking_number", "shipped_at", "created_at", "tag"], "title": "ShippedOrder", "type": "object"}, "OrderErrors": {"anyOf": [{"$ref": "#/components/schemas/OrderNotFoundError"}, {"$ref": "#/components/schemas/OrderNotAccessibleError"}], "title": "OrderErrors"}, "OrderNotAccessibleError": {"properties": {"tag": {"const": "OrderNotAccessibleError", "default": "OrderNotAccessibleError", "title": "Tag", "type": "string"}, "detail": {"title": "Detail", "type": "string"}, "id": {"title": "Id", "type": "integer"}}, "required": ["detail", "id", "tag"], "title": "OrderNotAccessibleError", "type": "object"}, "OrderNotFoundError": {"properties": {"tag": {"const": "OrderNotFoundError", "default": "OrderNotFoundError", "title": "Tag", "type": "string"}, "detail": {"title": "Detail", "type": "string"}, "id": {"title": "Id", "type": "integer"}}, "required": ["detail", "id", "tag"], "title": "OrderNotFoundError", "type": "object"}, "AvailableProduct": {"properties": {"tag": {"const": "available", "default": "available", "title": "Tag", "type": "string"}, "id": {"anyOf": [{"type": "integer"}, {"type": "null"}], "title": "ID"}, "name": {"maxLength": 255, "title": "Name", "type": "string"}, "description": {"title": "Description", "type": "string"}, "price": {"pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$", "title": "Price", "type": "string"}, "stock_count": {"default": 0, "title": "Stock Count", "type": "integer"}}, "required": ["name", "description", "price", "tag"], "title": "AvailableProduct", "type": "object"}, "OutOfStockProduct": {"properties": {"tag": {"const": "out_of_stock", "default": "out_of_stock", "title": "Tag", "type": "string"}, "id": {"anyOf": [{"type": "integer"}, {"type": "null"}], "title": "ID"}, "name": {"maxLength": 255, "title": "Name", "type": "string"}, "description": {"title": "Description", "type": "string"}, "price": {"pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$", "title": "Price", "type": "string"}}, "required": ["name", "description", "price", "tag"], "title": "OutOfStockProduct", "type": "object"}, "ProductResponse": {"anyOf": [{"$ref": "#/components/schemas/AvailableProduct"}, {"$ref": "#/components/schemas/OutOfStockProduct"}], "title": "ProductResponse"}, "ProductErrors": {"anyOf": [{"$ref": "#/components/schemas/ProductNotFoundError"}, {"$ref": "#/components/schemas/ProductHiddenError"}], "title": "ProductErrors"}, "ProductHiddenError": {"properties": {"tag": {"const": "ProductHiddenError", "default": "ProductHiddenError", "title": "Tag", "type": "string"}, "detail": {"title": "Detail", "type": "string"}, "id": {"title": "Id", "type": "integer"}}, "required": ["detail", "id", "tag"], "title": "ProductHiddenError", "type": "object"}, "ProductNotFoundError": {"properties": {"tag": {"const": "ProductNotFoundError", "default": "ProductNotFoundError", "title": "Tag", "type": "string"}, "detail": {"title": "Detail", "type": "string"}, "id": {"title": "Id", "type": "integer"}}, "required": ["detail", "id", "tag"], "title": "ProductNotFoundError", "type": "object"}}, "securitySchemes": {"JWTAuth": {"type": "http", "scheme": "bearer"}}}, "servers": []}